Factor out the verification into a single routine.
Different sanity checks were done all over the place. For instance,
the invocation parsing set a few parameters in the TransformationManager.

This unneccessary "Builder" design pattern makes things terrible to debug. We
should try to avoid it where possible.

This patch introduces the missing concept of InvocationOptions. They are passed
to the TransformationManager. Also initializeCompilerInstance and verify only
make sense in the context of doTransformation. Sink them there.