In cases we don't have all include paths we may not see the definition.
This fixes assertions in creduce. One might think that it is okay to proceed
with that transformation because the compiler we are testing will have it.
This might be true, however as part of the transformation we create:
static some_type s. We have no way to check for a default ctor thus making the
transformation yield false positive. That can be very harmful when run in bisect
mode.