RemoveDecl pass can try to remove a decl which was already removed.
This happens when a decl encloses another one. If the enclosing decl is deleted
and clang_delta tries to remove an inner decl it asserts. This happens in bisect
mode.

This patch teaches remove-decl to figure out if a candidate for removal encloses
another one and replaces it in the candidate list.