Change implementation of remove-decl.
One of the issues with the pass in binsearch mode was that if a decl A contains
decl B, A gets removed and creduce requests removing of B we get a hard to debug
crash. Now we track if the removed decl contains the requested to be removed one
and skip the removal, avoiding the crash.